Add probe function
This function describes how to respond to agents action. Produces one of
4 results: ILLEGAL if the move was not allowed (such as shooting outside
of the field), MISS if the shot missed, HIT if a ship was hit and SUNK
if a ship sunk because of the shot.

For now if the shot is repeated the
behaviour of the function is undefined.