Separate the gameplay from the server
With this update one can make two agents play the game on a local
machine, nothing regading networking. For a remote game one can use
`RemoteAgent` class, which forwards all the info/requests it gets down a
socket.

Note: playing a game can affect inner state of an agent so it is highly
recommended to not put 1 agent instance into 2 games at once.